pickle out the nltk model and refined simple_server
