+++ launch.py
@@ -1,4 +1,5 @@
-from static_frontend.simpleServer import app
+from backend.classifiers.nltktest import NLTKTest
+from backend.simple_server import app
 
 if __name__ == '__main__':
     app.run()
